requirements.txt
backcall==0.1.0
decorator==4.3.0
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
numpy==1.14.3
opencv-python==3.4.1.15
parso==0.2.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytesseract==0.2.0
simplegeneric==0.8.1
six==1.11.0
tesseract==0.1.3
traitlets==4.3.2
wcwidth==0.1.7
